chore: fikser  filstruktur for Table + props i Storybook docs
