+++ packages/dds-components/src/components/DetailList/DetailList.tsx
@@ -7,7 +7,7 @@ import {
 } from '../../types';
 import { cn } from '../../utils';
 import { type Breakpoint } from '../layout';
-import tableStyles from '../Table/normal/Table.module.css';
+import tableStyles from '../Table/Table.module.css';
 import tgStyles from '../Typography/typographyStyles.module.css';
 
 export const DETAIL_LIST_SIZES = createSizes('small', 'medium', 'large');
