--- packages/dds-components/src/components/Table/collapsible/index.ts
@@ -1,6 +1,0 @@
-import {
-  CollapsibleTable,
-  type CollapsibleTableProps,
-} from './CollapsibleTable';
-
-export { CollapsibleTable, type CollapsibleTableProps };
