+++ packages/dds-components/src/components/Table/index.ts
@@ -1,23 +1,58 @@
-export { CollapsibleTable, type CollapsibleTableProps } from './collapsible';
-export {
-  Table,
-  TableBody,
-  TableCell,
-  TableFoot,
-  TableHead,
-  TableRow,
-  TableSortCell,
-  TableWrapper,
-  type TableCellProps,
-  type TableSize,
-  type TableProps,
-  type TableRowProps,
-  type TableRowType,
-  type TableHeadProps,
-  type TableBodyProps,
-  type TableFootProps,
-  type TableSortCellProps,
-  type SortOrder,
+import { Body, type TableBodyProps } from './Body';
+import {
+  Cell,
   type TableCellLayout,
+  type TableCellProps,
   type TableCellType,
-} from './normal';
+} from './Cell';
+import { Foot, type TableFootProps } from './Foot';
+import { Head, type TableHeadProps } from './Head';
+import { Row, type TableRowProps, type TableRowType } from './Row';
+import { SortCell, type SortOrder, type TableSortCellProps } from './SortCell';
+import { Table as BaseTable, type TableProps, type TableSize } from './Table';
+import { TableWrapper } from './TableWrapper';
+
+type TableCompoundProps = typeof BaseTable & {
+  Wrapper: typeof TableWrapper;
+  Head: typeof Head;
+  Body: typeof Body;
+  Foot: typeof Foot;
+  Row: typeof Row;
+  Cell: typeof Cell;
+  SortCell: typeof SortCell;
+};
+
+const Table = BaseTable as TableCompoundProps;
+
+Table.Wrapper = TableWrapper;
+Table.Head = Head;
+Table.Body = Body;
+Table.Cell = Cell;
+Table.SortCell = SortCell;
+Table.Row = Row;
+Table.Foot = Foot;
+
+export { Table };
+
+export type {
+  TableCellProps,
+  TableSize,
+  TableProps,
+  TableRowProps,
+  TableRowType,
+  TableHeadProps,
+  TableBodyProps,
+  TableFootProps,
+  TableSortCellProps,
+  SortOrder,
+  TableCellLayout,
+  TableCellType,
+};
+
+export { TableWrapper };
+export { Head as TableHead };
+export { Body as TableBody };
+export { Cell as TableCell };
+export { SortCell as TableSortCell };
+export { Row as TableRow };
+export { Foot as TableFoot };
