--- packages/dds-components/src/components/Table/normal/Table.mdx
@@ -1,197 +1,0 @@
-import {
-  Canvas,
-  Story,
-  Meta,
-  ArgTypes,
-  Controls,
-} from '@storybook/addon-docs/blocks';
-import { TableRow } from '.';
-import { Table, CollapsibleTable } from '../';
-import { Tag } from '../../Tag';
-import {
-  Source,
-  ComponentLinkRow,
-} from '@norges-domstoler/storybook-components';
-import * as TableStories from './Table.stories';
-import meta from './Table.stories';
-import * as CollapsibleTableStories from '../collapsible/CollapsibleTable.stories';
-
-<Meta of={meta} />
-
-# Table
-
-<ComponentLinkRow
-  docsHref="559cc8-table"
-  figmaHref="node-id=1347%3A0"
-  githubHref={{ folder: 'components', comp: 'Table' }}
-  storybookHref={{ folder: 'components', comp: 'table' }}
-/>
-
-Tabeller bygges ved å bruke styled versjoner av native HTML elementer. Disse komponentene er:
-
-- **`<Table.Wrapper>`** - wrapper for tabell.
-- **`<Table>`** - selve tabellen.
-- **Subkomponenter:** `<Table.Head>`, `<Table.Body>`, `<Table.Foot>`, `<Table.Row>`, `<Table.Cell>`, `<Table.SortCell>`.
-- **`<CollapsibleTable>`** - variant av tabell som kan kollapses, med tilhørende rad-komponent: `<CollapsibleTable.Row>`.
-
-Enkelte subkomponenter har props for å sørge for riktig styling og oppførsel, mens andre har kun native props.
-
-Ikke-compound komponenter er også støttet: `<TableWrapper>`, `<TableHead>`, `<TableBody>`, `<TableFoot>`, `<TableRow>`, `<TableCell>`, `<TableSortCell>`, `<CollapsibleTableRow>`.
-
-## Varianter
-
-### Tabell
-
-<Canvas of={TableStories.Preview} />
-
-### Kollapsibel tabell
-
-<Canvas of={CollapsibleTableStories.SingleDefiningColumn} />
-
-## Props
-
-### Table.Wrapper
-
-En wrapper som sørger for at tabellen får scrolling når den ikke får plass i bredden.
-
-Støtter alle native HTML-attributter som er en del av `HTMLAttributes<HTMLDivElement>`-interface.
-
-### Table
-
-Tilsvarer `<table>`.
-
-<Canvas of={TableStories.Preview} />
-<Controls of={TableStories.Preview} />
-
-I tillegg støttes alle native HTML-attributter som er en del av `HTMLAttributes<HTMLTableElement>`-interface.
-
-### Table.Body
-
-Tilsvarer `<tbody>`.
-
-Støtter alle native HTML-attributter som er en del av `HTMLAttributes<HTMLTableSectionElement>`-interface.
-
-### Table.Head
-
-Tilsvarer `<thead>`.
-
-Støtter alle native HTML-attributter som er en del av `HTMLAttributes<HTMLTableSectionElement>`-interface.
-
-### Table.Foot
-
-Tilsvarer `<tfoot>`.
-
-Støtter alle native HTML-attributter som er en del av `HTMLAttributes<HTMLTableSectionElement>`-interface.
-
-### Table.Row
-
-Tilsvarer `<tr>`.
-
-<ArgTypes of={TableRow} />
-
-Støtter alle native HTML-attributter som er en del av `HTMLAttributes<HTMLTableSectionElement>`-interface.
-
-### Table.Cell
-
-Tilsvarer `<th>` eller `<td>`.
-
-<ArgTypes of={Table.Cell} />
-
-I tillegg støttes alle native HTML-attributter som er en del av `HTMLAttributes<HTMLTableDataCellElement>` og `HTMLAttributes<HTMLTableHeaderCellElement>`-interface.
-
-### Table.SortCell
-
-Et styled `<Cell type='head'>`-element som brukes som `<th>` i sorteringskolonner. **OBS!** komponenten stiller kun med styling og ikoner, sortering implementeres i de respektive løsningene.
-
-<ArgTypes of={Table.SortCell} />
-
-I tillegg støttes `<Cell>`-props, bortsett fra `type` da den er satt til `'head'`.
-
-### CollapsibleTable
-
-En variant av `<Table>` som kan kollapse deler av hver rad.
-Konsumenten må bestemme en eller flere såkalte definerende kolonner - kolonner som vises til enhver tid som i vanlig tabell, mens reseterende kolonner kan kollapse.
-Det legges en kolonne til høyre som inneholder en knapp til utviding av raden.
-Resterende innhold blir kollapset og vises ved hjelp av `<DescriptionList>`.
-
-Det må brukes `<CollapsibleTable.Row>` istedenfor `<Table.Row>` for riktig oppførsel.
-Kollapset tilstand brukes som mobilversjon.
-
-Rekkefølgen på definerende kolonner følger rekkefølgen i `definingColumnIndex`.
-
-<Canvas of={CollapsibleTableStories.SingleDefiningColumn} />
-<Controls of={CollapsibleTableStories.SingleDefiningColumn} />
-
-### CollapsibleTable.Row
-
-En variant av `<Table.Row>` som brukes med `<CollapsibleTable>`. Den håndterer kollapsing.
-
-<ArgTypes of={CollapsibleTable.Row} />
-
-## Bruk
-
-<Source
-  code={`import { Table, CollapsibleTable } from '@norges-domstoler/dds-components';
-
-// Normal
-
-<Table.Wrapper>
-  <Table>
-    <Table.Head>
-      <Table.Row>
-        <Table.Cell>Navn</Table.Cell>
-        <Table.Cell>Rolle</Table.Cell>
-      </Table.Row>
-    </Table.Head>
-    <Table.Body>
-      <Table.Row>
-        <Table.Cell>Sandra Lovsetter</Table.Cell>
-        <Table.Cell>Admin</Table.Cell>
-      </Table.Row>
-      <Table.Row>
-        <Table.Cell>Marie Bjerke</Table.Cell>
-        <Table.Cell>Admin</Table.Cell>
-      </Table.Row>
-    </Table.Body>
-  </Table>
-</Table.Wrapper>
-
-// Collapsible
-
-const headerValues = [
-{
-key: 'navn',
-content: 'Navn',
-},
-{
-key: 'rolle',
-content: 'Rolle',
-}
-]
-
-<CollapsibleTable>
-  <Table.Head>
-    <CollapsibleTable.Row headerValues={headerValues}>
-      <Table.Cell>{headerValues[0].content}</Table.Cell>
-      <Table.Cell>{headerValues[1].content}</Table.Cell>
-    </CollapsibleTable.Row>
-  </Table.Head>
-  <Table.Body>
-    <CollapsibleTable.Row>
-      <Table.Cell>Sandra Lovsetter</Table.Cell>
-      <Table.Cell>Admin</Table.Cell>
-    </CollapsibleTable.Row>
-    <CollapsibleTable.Row>
-      <Table.Cell>Marie Bjerke</Table.Cell>
-      <Table.Cell>Admin</Table.Cell>
-    </CollapsibleTable.Row>
-  </Table.Body>
-</CollapsibleTable>
-`} />
-
-## Retningslinjer
-
-- Bruk `<Table.Foot>` som footer i tabellen ved summering av verdier i en kolonne eller lignende.
-- Hele raden skal ikke være klikkbar. Hvis det trengs en handling på en rad skal det brukes lenker eller knapper i selve raden:
-
-<Canvas of={TableStories.InteractiveContent} />
